requirements.txt
Django==5.2.4
djangorestframework==3.14.0
openai==1.3.7
python-dotenv==1.0.0
django-cors-headers==4.3.1 